\name{MR.LDP-package}
\alias{MR.LDP-package}
\alias{MR.LDP}
\docType{package}
\title{
\packageTitle{MR.LDP}
}
\description{
\packageDescription{MR.LDP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MR.LDP}
\packageIndices{MR.LDP}
}
\author{
\packageAuthor{MR.LDP}

Maintainer: \packageMaintainer{MR.LDP}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
