\name{MR.LDP-package}
\alias{MR.LDP-package}
\alias{MR.LDP}
\docType{package}
\title{
\packageTitle{MR.LDP}
}
\description{
\packageDescription{MR.LDP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MR.LDP}
\packageIndices{MR.LDP}
An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MR.LDP}

Maintainer: \packageMaintainer{MR.LDP}
}
\references{
Literature or other references for background information
}
% Optionally other standard keywords, one per line, from file KEYWORDS in the R
documentation directory
\keyword{ package }
\seealso{
Optional links to other man pages, e.g.
\code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{
simple examples of the most important functions
}
